Exclude example files from published package
During our dependency reviews we discovered that the muda crate as uploaded to crates.io includes example data (a png)  that are not used by the rust code. This increases the package size and makes it harder to review the source code as it now includes binary data.

This commit explicitly includes only relevant files. This helps to reduce the package size

Before: 54 files, 482.0KiB (105.0KiB compressed)
After: 39 files, 416.9KiB (87.1KiB compressed)

Overall based on the reduction in the compressed package size and the current download numbers that results to a traffic reduction for crates.io of around 12GB / Month.